+++ Block/Zone/Summary.php
@@ -38,12 +38,14 @@ class Summary extends AbstractZone
     /**
      * @inheritdoc
      */
-    public function addToSummary(string $sectionName, string $key, mixed $value): void
+    public function addToSummary(string $sectionName, string $key, mixed $value): self
     {
         if (is_array($value) && array_key_exists('has_warning', $value) && $value['has_warning']) {
             $this->hasWarning();
         }
 
         $this->summary[$sectionName][$key] = $value;
+
+        return $this;
     }
 }